ruff_python_formatter: move docstring handling to sub-module
Suggested here: https://github.com/astral-sh/ruff/pull/8811#discussion_r1401169531
